+++ .eslintrc.js
@@ -7,6 +7,7 @@ module.exports = {
     'ecmaVersion': 6,
     'ecmaFeatures': {
         'jsx': true,
+        'experimentalObjectRestSpread': true
       }
     },
     'env': {